requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0